pyproject.toml
[project]
name = "smartoj-backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2.0",
    "aiosmtplib>=4.0.1",
    "elasticsearch[async]==8.18",
    "fastapi[standard]>=0.115.12",
    "filetype>=1.2.0",
    "httpx-socks>=0.10.0",
    "minio>=7.2.15",
    "mysql-replication==0.46",
    "python-socks>=2.7.1",
    "sqlmodel>=0.0.24",
    "taskiq>=0.11.17",
    "taskiq-aio-pika>=0.4.2",
    "taskiq-redis>=1.0.8",
    "user-agents>=2.2.0",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
